Deduplicate resolve functions
